# UCP/PROTO/RNDV: Use rma_bw_lanes for lanes sorting