--- a/src/ucp/proto/proto_common.c
+++ b/src/ucp/proto/proto_common.c
@@ -449,11 +449,45 @@ err_deref_perf_node:
     return status;
 }
 
+/*
+ * TODO: This is a quickfix, needed to select lanes for multi-lane RNDV
+ * protocol in the order of rma_bw_lanes (RMA_BW lanes sorted by score).
+ * The proper solution is to have a generic mechanism to sort lanes based on
+ * the calculated performance, implemented in proto_multi.
+ * This function should be removed once the proper solution is implemented.
+ */
+static inline ucp_lane_index_t
+ucp_proto_common_lanes_iter(const ucp_ep_config_key_t *ep_config_key,
+                            ucp_lane_map_t lane_map, ucp_lane_type_t lane_type,
+                            ucp_lane_index_t start, ucp_lane_index_t *lane)
+{
+    if (start >= UCP_MAX_LANES) {
+        return UCP_MAX_LANES;
+    }
+
+    if (lane_type == UCP_LANE_TYPE_RMA_BW) {
+        for (; start < ep_config_key->num_lanes; ++start) {
+            *lane = ep_config_key->rma_bw_lanes[start];
+            if ((*lane == UCP_NULL_LANE) || (lane_map & UCS_BIT(*lane))) {
+                break;
+            }
+        }
+        return start;
+    }
+
+    /*
+     * By default iterate over all lanes in lane_map
+     * Reset lane_map bits below start position, then find first bit set
+     */
+    lane_map &= ~((1ULL << start) - 1);
+    *lane     = ucs_ffs64_safe(lane_map);
+    return *lane;
+}
+
 ucp_lane_index_t
 ucp_proto_common_find_lanes(const ucp_proto_init_params_t *params,
-                            uct_ep_operation_t memtype_op, unsigned flags,
-                            ptrdiff_t max_iov_offs, size_t min_iov,
-                            ucp_lane_type_t lane_type,
+                            unsigned flags, ptrdiff_t max_iov_offs,
+                            size_t min_iov, ucp_lane_type_t lane_type,
                             ucs_memory_type_t reg_mem_type,
                             uint64_t tl_cap_flags, ucp_lane_index_t max_lanes,
                             ucp_lane_map_t exclude_map, ucp_lane_index_t *lanes)
@@ -464,7 +498,7 @@ ucp_proto_common_find_lanes(const ucp_proto_init_params_t *params,
     const ucp_rkey_config_key_t *rkey_config_key = params->rkey_config_key;
     const ucp_proto_select_param_t *select_param = params->select_param;
     const uct_iface_attr_t *iface_attr;
-    ucp_lane_index_t lane, num_lanes;
+    ucp_lane_index_t lane, num_lanes, i;
     const uct_md_attr_v2_t *md_attr;
     const uct_component_attr_t *cmpt_attr;
     ucp_rsc_index_t rsc_index;
@@ -497,7 +531,12 @@ ucp_proto_common_find_lanes(const ucp_proto_init_params_t *params,
     }
 
     lane_map = UCS_MASK(ep_config_key->num_lanes) & ~exclude_map;
-    ucs_for_each_bit(lane, lane_map) {
+    lane     = 0;
+    for (i = ucp_proto_common_lanes_iter(ep_config_key, lane_map, lane_type,
+                                         0, &lane);
+         (i < ep_config_key->num_lanes) && (lane != UCP_NULL_LANE);
+         i = ucp_proto_common_lanes_iter(ep_config_key, lane_map, lane_type,
+                                         i + 1, &lane)) {
         if (num_lanes >= max_lanes) {
             break;
         }
@@ -655,10 +694,9 @@ ucp_lane_index_t ucp_proto_common_find_lanes_with_min_frag(
     size_t tl_min_frag, tl_max_frag;
 
     num_lanes = ucp_proto_common_find_lanes(
-                   &params->super, params->memtype_op, params->flags,
-                   params->max_iov_offs, params->min_iov, lane_type,
-                   params->reg_mem_info.type, tl_cap_flags, max_lanes,
-                   exclude_map, lanes);
+                   &params->super, params->flags, params->max_iov_offs,
+                   params->min_iov, lane_type, params->reg_mem_info.type,
+                   tl_cap_flags, max_lanes, exclude_map, lanes);
 
     num_valid_lanes = 0;
     for (lane_index = 0; lane_index < num_lanes; ++lane_index) {