## What?
This is fix for [RM#4177809](https://redmine.mellanox.com/issues/4177809).
The root cause of the issue is sorting lanes by BW introduced in https://github.com/openucx/ucx/pull/10264
That PR solves one issue, but introduces another one
- It solves issue when we always choose first AM lane for RNDV, even if it's slower
- It introduces the issue that we may choose several paths of the same device, even though it's not optimal (must prefer another device instead)

This is a quick fix, the ideal solution would be to do lane sorting in proto_multi after performance calculation, taking into account multiple paths and performance boost of extra paths

## How?
- Revert lane sorting by BW (https://github.com/openucx/ucx/pull/10264)
- Rely on rma_bw_lanes array for finding RMA_BW lanes for RNDV
- Tested with latest version of protocol mock (https://github.com/openucx/ucx/pull/10369)

Tested on ganon:
```mpirun -np 32 --map-by ppr:16:node -x LD_PRELOAD=/home/iyastrebov/storage/ucx/bld-devel/lib/libucp.so:/home/iyastrebov/storage/ucx/bld-devel/lib/libuct.so:/home/iyastrebov/storage/ucx/bld-devel/lib/libucs.so --mca coll_ucc_enable 0 -x UCX_MULTI_PATH_RATIO=0.9 /home/mmagazinnik/PerfX/OSU_40341/osu_mbw_mr D D```

BEFORE:
```
2100..inf | (?) rendezvous zero-copy read from remote | rc_mlx5/mlx5_0:1 50% on path0 and 50% on path1

# Size                  MB/s        Messages/s
131072              49450.16         377274.79
262144              49653.20         189411.92
524288              49617.67          94638.20
1048576             49591.96          47294.58
2097152             49581.84          23642.46
4194304             49572.40          11818.98
```
AFTER:
```
2100..inf | (?) rendezvous zero-copy read from remote | 80% on rc_mlx5/mlx5_0:1/path0 and 20% on rc_mlx5/mlx5_1:1

131072              60445.56         461163.01                                                                   
262144              60461.71         230643.11                                                                   
524288              61062.24         116466.98                                                                   
1048576             61251.31          58413.80                                                                   
2097152             60680.78          28934.85                                                                   
4194304             60366.36          14392.46                                                                   
```